+++ debug.log
@@ -1018,3 +1018,93 @@
 [0203/200844.516:ERROR:crash_report_database_win.cc(469)] failed to stat report
 [0203/200844.516:ERROR:crash_report_database_win.cc(469)] failed to stat report
 [0203/200844.516:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0203/222446.271:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0203/222446.271:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0203/222446.271:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0203/222446.272:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0203/222446.272:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0203/222446.272:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0203/222446.272:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0203/222446.272:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0203/222446.272:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0203/222446.272:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/000254.750:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/000254.751:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/000254.751:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/000254.751:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/000254.751:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/000254.751:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/000254.751:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/000254.751:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/000254.751:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/000254.751:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/165640.434:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/165640.450:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/165640.450:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/165640.450:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/165640.450:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/165640.451:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/165640.451:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/165640.451:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/165640.451:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/165640.451:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/202313.758:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/202313.758:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/202313.758:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/202313.758:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/202313.758:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/202313.758:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/202313.758:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/202313.758:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/202313.758:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/202313.758:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/210250.097:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/210250.097:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/210250.097:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/210250.098:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/210250.098:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/210250.098:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/210250.098:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/210250.098:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/210250.098:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/210250.098:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/211850.929:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/211850.929:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/211850.929:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/211850.929:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/211850.929:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/211850.929:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/211850.929:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/211850.929:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/211850.929:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/211850.929:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/221101.963:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/221101.964:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/221101.964:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/221101.964:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/221101.964:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/221101.964:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/221101.964:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/221101.964:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/221101.964:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/221101.964:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/221223.686:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/221223.686:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/221223.686:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/221223.686:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/221223.686:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/221223.686:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/221223.686:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/221223.687:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/221223.687:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/221223.687:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/223008.797:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/223008.797:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/223008.797:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/223008.797:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/223008.797:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/223008.797:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/223008.797:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/223008.797:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/223008.797:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0204/223008.797:ERROR:crash_report_database_win.cc(469)] failed to stat report
