--- src/index.css
@@ -1,18 +1,0 @@
-@import url(/src/pages/vendor/normalize.css);
-@import url(/src/pages/vendor/fonts.css);
-@import url(/src/blocks/root/root.css);
-@import url(/src/blocks/header/header.css);
-@import url(/src/blocks/logo/logo.css);
-@import url(/src/blocks/profile/profile.css);
-@import url(/src/blocks/user-info/user-info.css);
-@import url(/src/blocks/button/button.css);
-@import url(/src/blocks/places-list/places-list.css);
-@import url(/src/blocks/place-card/place-card.css);
-@import url(/src/blocks/popup/popup.css);
-@import url(/src/blocks/popup-profile/popup-profile.css);
-@import url(/src/blocks/popup-image/popup-image.css);
-
-
-body {
-  background-color: #000;
-}
