+++ src/modules/Card.js
@@ -1,3 +1,5 @@
+export const placesList = document.querySelector('.places-list');
+
 class Card {
   constructor(link, name) {
     this.cardElement = this.create(link, name);
@@ -51,3 +53,27 @@ class Card {
     event.stopPropagation();
   }
 }
+
+class CardList {
+  constructor(container, cards) {
+    this.container = container;
+    this.cards = cards;
+  }
+  //Добавление карточки 
+  addCard(link, name) {
+    const { cardElement } = new Card(link, name);
+
+  }
+  //Рендер первоначальных карточек
+  render() {
+    const newArray = this.cards.map(function (item) {
+      const { link, name } = item;
+      return new Card(link, name);
+    });
+  }
+}
+
+
+
+export {Card};
+export {CardList};
