--- src/modules/initialCards.js
@@ -1,9 +1,0 @@
-function initialCards(){
-    api.getInitialCards()
-        .then((cards) => {
-        const cardList = new CardList(placesList, cards)
-        cardList.render(cards)
-})
-}
-
-initialCards();
