--- src/modules/takeUserInfo.js
@@ -1,9 +1,0 @@
-function takeUserInfo(){
-    
-  api.getUserInfo().then((res) =>{
-      userName.textContent = res.name; 
-      userJob.textContent = res.about;     
-    })
-}
-
-takeUserInfo();
